cros-codecs: Support DecodedFormat::YV12
This CL adds support for a new DecodedFormat::YV12 which is like I420
but with the U and V planes swapped. This is needed for encoding. The
i420_to_nv12() function is re-used to implement the conversion from YV12
to NV12.

BUG=b:389993558
TEST=VtsHalMediaC2V1_0TargetVideoEncTest on brya

Change-Id: I8c261a741f5f2e47066fe215f6ba934ec0a082b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6674644
Auto-Submit: Bijan Choobineh‎ <bchoobineh@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@google.com>
Tested-by: Bijan Choobineh‎ <bchoobineh@chromium.org>
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>