cros-codecs: AV1: Reset last frame render resolution on sequence header
We keep track of the render resolution of the last frame in order to
process "frame size override" style DRCs, but we weren't resetting this
value upon receiving a new sequence header. This was breaking "sequence
header" style DRCs because we were using the coded size parsed from the
sequence header but the render resolution of the frame before the
sequence header. If the post-DRC resolution was smaller than the pre-DRC
resolution, then this results in a coded resolution smaller than the
render resolution, which causes a crash.

BUG=b:412861829
TEST=atest
android.mediav2.cts.AdaptivePlaybackTest#testAdaptivePlayback

Change-Id: Icf1a4771e6802e50901d3a73472b0b69d40a176a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6674641
Reviewed-by: Hang Nguyen <hnt@google.com>
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>
Auto-Submit: Bijan Choobineh‎ <bchoobineh@chromium.org>
Tested-by: Bijan Choobineh‎ <bchoobineh@chromium.org>