cros-codecs: h264: Don't reject bitstreams that have no NALUs
This CL allows the stateless H.264 decoder to move on after getting a
bitstream without NALUs instead of failing. This case was observed on a
bitstream that had a P frame followed by a bunch of 0s. Regardless of
whether that's a bug on the encoder or not, not rejecting the bitstream
outright makes the decoder more resilient in the wild.

BUG=b:414856492
TEST=play the previously problematic .mp4 on brya
TEST=VtsHalMediaC2V1_0TargetVideoDecTest on brya

Change-Id: I1ab8fef38ab3d4fd4d14cf7898ea7a20b01bc318
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6674643
Auto-Submit: Bijan Choobineh‎ <bchoobineh@chromium.org>
Reviewed-by: Hang Nguyen <hnt@google.com>
Reviewed-by: Andres Calderon Jaramillo <andrescj@google.com>
Tested-by: Bijan Choobineh‎ <bchoobineh@chromium.org>
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>