cros-disks: Enhance the drivefs SECCOMP policies
Add some potentially missing syscalls to the drivefs SECCOMP policies.

If a syscall is present in an architecture, ensure that it is present in
all the 3 supported architectures (AMD-64, ARM-64 and ARM-32).

If a syscall has a 32-bit and a 64-bit variants in an architecture,
ensure that both variants are present.

If a file-related syscall has a "plain" and an "*at" variant (e.g. mkdir
and mkdirat), ensure that both variants are present.

Ensure that the SECCOMP policies are sorted and that the syscalls are
listed in the same order as in
https://www.chromium.org/chromium-os/developer-library/reference/linux-constants/syscalls/#cross-arch-numbers

BUG=b:424206022
TEST=Manual

Change-Id: I3c63a2ecf24f9752381cda7b8997352b37ad02c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6651665
Auto-Submit: François Degros <fdegros@chromium.org>
Tested-by: François Degros <fdegros@chromium.org>
Reviewed-by: Ben Reich <benreich@chromium.org>
Commit-Queue: François Degros <fdegros@chromium.org>