vm_tools: concierge: assume Parallels VMs are not running or registered
Now that PDfC was disabled everywhere users that have Parallels VMs on
their devices need to be able to remove them even though Parallels
dispatcher is not running and concierge is unable to verify that the VM
is not running. Since it is not possible for a Parallels VM to be
running simply remove the checks so that "vmc" can succeed.

BUG=b:420750686
TEST=Start with M138 build, create a Parallels VM, upgrade to M139, make
sure the VM image can be exported and/or deleted with "vmc" command.

Change-Id: Ib8607773a76b888a158ffa8c7ecc30b4a59c4667
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6657932
Commit-Queue: Junichi Uekawa <uekawa@chromium.org>
Auto-Submit: Dmitry Torokhov <dtor@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>