cros-codecs: c2_wrapper: Don't call the `error_cb` outside of the worker thread
Currently, all of the calls to the `error_cb` in the C2Wrapper are
simply to notify the client that a C2Wrapper method was called while in
an unexpected state. The problem is that this doesn't happen on the
correct thread: if the client has already been used in the worker thread
(e.g., to notify of completed work), a call to the `error_cb` in the
C2Wrapper would violate the thread affinity of the client (this is not
actually a safety issue currently as it could only manifest as a crash).

Fortunately, it looks like such calls are not necessary because the
C2Wrapper methods already return C2BadState when called while in the
wrong state, so the caller already gets a signal. Additionally, taking
a look at the start(), stop(), queue_nb(), flush_sm(), and drain_nb()
methods of Android's SimpleC2Component [1], it's observed that the error
callback of the C2ComponentListener is not invoked when those methods
are called while in the wrong state.

Therefore, this CL simply removes the calls to the `error_cb` in the
C2Wrapper which also allows us to remove `error_cb` as a member of
C2Wrapper.

[1]
https://android.googlesource.com/platform/frameworks/av/+/41ed5de29ceee6c4f0faa08187aa2cc61795a6eb/media/codec2/components/base/SimpleC2Component.cpp

BUG=b:426730251
TEST=VtsHalMediaC2V1_0TargetMasterTest, VtsHalMediaC2V1_0TargetComponentTest, VtsHalMediaC2V1_0TargetVideoDecTest, VtsHalMediaC2V1_0TargetVideoEncTest on brya while monitoring `adb logcat -T 1 DEBUG:* *:S`
TEST=atest ccdec_test on brya
TEST=atest ccenc_test on brya

Change-Id: Idd480cd0f428969aa1a9a61e35d6adf40b765a44
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6674647
Auto-Submit: Bijan Choobineh‎ <bchoobineh@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@google.com>
Tested-by: Bijan Choobineh‎ <bchoobineh@chromium.org>
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>