anx7625-ta: Add TA for controlling ANX7625 display port bridge
This TA exists for the REE to interact with the ANX7625 when the I2C
control has been placed in the TEE. The TA API functions allow for safe
register reading and writing. The code is also checked into a ChromeOS
vendor repository:
https://chrome-internal.googlesource.com/chromeos/vendor/anx7625-ta.

Analogix okayed moving it to platform, which makes it easier to develop.
No proprietary logic is present that does not already exist in the open
source driver.

It also implements a power status API to let the REE "turn on" the
ANX7625 device.

There is a bit of functionality that is missing. The integration with
the Mediatek I2C kernel TA is completely fictional. There are security
TODO's for register writing and power status setting that need to be
resolved before launch.

BUG=b:418014229
TEST=compiles

Change-Id: I43c2f70765051d89878d2adc720937b3cb843b85
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6668716
Commit-Queue: Nathan Hebert <nhebert@chromium.org>
Tested-by: Nathan Hebert <nhebert@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>