shill: DHCP: Make dbus call in dhcpcd hook run in background
This is to resolve the tiny deadlock (which lasts for ~3s) between
shill and dhcpcd10.

When dhcpcd10 is used and DHCP provisioning process fails, shill will
first send SIGALRM to dhcpcd to let it release the lease, and wait for
dhcpcd to exit. If dhcpcd doesn't exit after ~3s, shill will send
SIGKILL to kill it.

Note that this wait is implemented in
ProcessManager::StopProcessAndBlock() and is synchronous. This means
shill cannot receive any DHCP event when waiting for the process to
exit.

Meanwhile after dhcpcd receives SIGALRM, it will try to release the
lease and will exit. During this process, it will trigger several events
through dhcpcd hooks, and the event reasons include "STOP", "STOPPED",
and "NOCARRIER" (when dhcpcd is still being initialized). The hook
script sends these events to shill using "gdbus call", which is also
synchronous. However, shill cannot respond because it is still waiting
for the dhcpcd to exit! As a result, shill freezes for around 3s, until
it kills dhcpcd with SIGKILL.

This is unacceptable and causes tast test network.RoutingNoIP to fail.

To break the deadlock, we can make the gdbus call in the dhcpcd hook run
in background so that it does not block dhcpcd execution. Before this
change, the IP configuration in shill and patchpanel is already async,
so this change will not cause more trouble if the gdbus call is delayed
due to shill being busy.

BUG=b:344500617
TEST=tast run $DUT network.RoutingNoIP (with dhcpcd10)
TEST=tast run $DUT network.*dhcpcd10 network.DHCPPD*

Change-Id: I776690e9aab04827bcae5b4fb0b79633812513a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6671490
Commit-Queue: Zikai Chen <chenzikai@google.com>
Tested-by: Zikai Chen <chenzikai@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Auto-Submit: Zikai Chen <chenzikai@google.com>