system-proxy: Allow SIGSYS in outer minijail kill policy
This change modifies the outer minijail's seccomp policy to permit
`kill` syscalls with `SIGSYS` in addition to `SIGTERM`. This
prevents the outer process from crashing when the inner process
(the worker) triggers a `SIGSYS` due to a forbidden syscall, and
allows the outer process to handle the error gracefully.

BUG=b:427596184
TEST=CQ

Change-Id: I380eb487296ff103cc0010ea8631b048e38b9527
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6701218
Commit-Queue: Andreea Costinas <acostinas@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Andreea Costinas <acostinas@google.com>