cryptohome: Make pinweaver passwords the default
This feature is being gradually enabled in stable now, so we should be
able to flip the default to enabled.

This requires updating a few unit tests that were explicitly checking
the password behavior. This was expected, the behavior of these tests
was deliberate.

BUG=b:391427550
TEST=unit tests

Change-Id: I9902d9d5c4a38c23a3c85f11f1519f958be717bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6714940
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>