installer: Mitigate udevadm settle timeout
During reload_partitions, the call to `udevadm settle` can time out.
Reduce the timeout to 10 seconds and ignore the non-zero exit
to continue on.

Add logging to mark when we start and finish reloading the
partition information to further help triage this issue in
the future.

BUG=b:379085243
TEST=cros build-image --board=reven; cargo xtask test-install

Change-Id: I1b73169229761ba45f4e2ad6fd9d90f1fd8159ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6802480
Commit-Queue: Justin Ouellette <justinoue@google.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>
Tested-by: Justin Ouellette <justinoue@google.com>