hardware_verifier: Refresh Runtime HWID in verifier instead
To skip the component categories that don't exist in the hardware
verification spec while checking if Runtime HWID should be generated, we
need to parse the spec. The parsing logic is already in the verifier
implementation. To avoid duplicate code, this CL moves
`RefreshRuntimeHWID()` from `HwVerificationReportGetterImpl` to
`VerifierImpl`.

BUG=b:367899338
TEST=FEATURES=test emerge-${BOARD} hardware_verifier

Change-Id: Ia344ec95193951ee464f568d66ac0a31e1764b77
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6810626
Commit-Queue: Allen Shih <allenshihmc@chromium.org>
Tested-by: Allen Shih <allenshihmc@chromium.org>
Reviewed-by: Yu-An Wang <wyuang@google.com>