missive: Increase initial delay for uploads
After non-periodic queue initiates upload, check back and retry after
3 sec instead of 1 sec. Chrome uploader is doing an exponential backoff
and was supposed to soften the impact of 1 sec retry (to possibly update
the cache but not send anything to the server); however, according to
the bug, it is insufficient or not functioning properly (heuristics of
duplicate events detection might be not aggressive enough).

The change should not have a adverse impact - it affects only events
that are not delivered soon enough.

BUG=b:435234655
TEST=cros_sdk cros_workon_make --board=${BOARD} missive --test
TEST=Manual VM test

Change-Id: Icd3289b7be133b711273a3c90bd7ef75caef2990
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6825361
Reviewed-by: Vignesh Shenvi <vshenvi@google.com>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>
Auto-Submit: Leonid Baraz <lbaraz@chromium.org>