login: Migrate hardware filter to Chrome Setup
The hardware filter logic moved from the Chromium codebase to ChromeOS
in this change. This change is necessary because ChromeOS Flex devices
require a hardware filter check to determine if ARC can be enabled.

Previously, the hardware filter logic was in the Chromium codebase,
which couldn't guarantee that the hardware check result would be
available before IsArcAvailable() was called from different parts of
the code.

By migrating the hardware filter to login_manager, we ensure the check
is performed early. In this code change, the hardware result will be
passed as a command-line flag when starting Chrome.

BUG=b:441376146
TEST=cros_workon_make --board=reven chromeos-base/chromeos-login --test

Cq-Depend: chromium:6898303
Change-Id: I6f55267e330c30e72c040292570ff1238db653d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6872462
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: TingHao Lin <tinghaolin@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Tested-by: TingHao Lin <tinghaolin@google.com>