vm_tools: Make ARC root path a shared mountpoint
This change configures the ARC root path as a shared mountpoint for the
vm_concierge service. On reven boards, the ARCVM image needs to be
downloaded from DLC at runtime, and then the DLC package is bind-mounted
to the ARC root path after the download is complete. If vm_concierge
starts before this installation and bind mount, the vm_concierge running
in minijail cannot see the change to the ARC root path. By making the
ARC root path a shared mountpoint, it can observe the bind mount changes
from outside the minijail environment.

BUG=b:443987476
TEST=emerge-reven vm_host_tools
TEST=Deploy the change and confirm that vm_concierge can see the ARC root path change

Change-Id: Ia2a478847f3b6a42bbb76611df9385b1d9c6ac54
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6932323
Reviewed-by: Lepton Wu <lepton@chromium.org>
Tested-by: TingHao Lin <tinghaolin@google.com>
Reviewed-by: maciek swiech <drmasquatch@google.com>
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: TingHao Lin <tinghaolin@google.com>