flex_hwis: Remove JSON fwupd history parsing
The JSON parsing will be replaced with dbus parsing in a later
commit. This switch to dbus parsing is being made to avoid having to
invoke the fwupdmgr executable, which would require making the minijail
much more permissive. The reason for removing the JSON parsing now, at
the beginning of the CL chain, is that it will allow removing some uses
of unique_ptr in the following commit, which makes the fwupd types
easier to copy.

Nothing uses the JSON parsing outside of tests, so removing it will not
affect functionality.

BUG=b:392122853
TEST=FEATURES=test emerge-reven chromeos-base/flex_hwis

Change-Id: I0218c80f2064117107ee390700f91efc12f2c8e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6956021
Reviewed-by: TingHao Lin <tinghaolin@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>