sepolicy: Fix flex_device_metrics directory label
The previous policy was attempting to label a file in the directory, but
skipped labeling the directory itself. This caused a problem because the
directory is created/updated by tmpfiles, which fails if the directory
doesn't have a file context entry.

Change the label name to cros_run_flex_device_metrics, and apply it to
both the directory and its contents.

BUG=b:392122853
TEST=emerge-reven chromeos-base/flex_hwis selinux-policy
TEST=build image, run on DUT
TEST=Check directory label: ls -lZa /run/flex_device_metrics

Change-Id: I1bca3d07790b1ffa933d2596128d32809f1a5e42
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6954997
Reviewed-by: Allen Webb <allenwebb@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>