flex_hwis: Combine fwupd timestamp functions
Merge GetFwupMetricTimestamp and RecordFwupMetricTimestamp into a single
function: GetAndUpdateFwupMetricTimestamp.

The logic around reading and updating the timestamp is a bit subtle (we
never want to forget to _write_ the timestamp, even if some error occurs
when reading it), so combine them into one function that can be tested
as a whole, and rework the related unit tests accordingly.

BUG=b:392122853
TEST=FEATURES=test emerge-reven chromeos-base/flex_hwis

Change-Id: If272da2005ea2e03c8dd6be0f578c8f6486d22d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6962497
Reviewed-by: TingHao Lin <tinghaolin@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>