chromeos-config: cros_config_host: Validate power button is defined once
A power button is either on the keyboard (keyboard.power_button is
present) or elsewhere (on the edge of the screen or keyboard plane),
but not both.

Raise an exception when the power button is defined twice.

BUG=b:445702150
TEST=Check on a faulty board (rex/rex) that an issue is raised.
Check it works fine on valid board (skywalker/anakin)

Cq-Depend: chromium:6962960
Change-Id: I29596576b0dfc0862a8dabf5dead580fc4aefb91
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6963805
Reviewed-by: Aamir Bohra <aamirbohra@google.com>
Auto-Submit: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>