flex_hwis: Use fwupd history's modification time
Use the `modified` field instead of the `created` field. The created
field is only changed when inserting an entry into the history, whereas
the modified field is updated any time the entry is modified. This is
signficant because a device may be updated multiple times
(e.g. installing different versions of firmware). Each of those events
should send a UMA.

BUG=b:392122853
TEST=FEATURES=test emerge-reven chromeos-base/flex_hwis
TEST=Deploy to dut, verify a UMA is sent after each update to a device

Change-Id: I8368730f4681d3adf7f77f161fd9141e44f295fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6966219
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: TingHao Lin <tinghaolin@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>