power: suspend_stress_test: Ignore waived sighting fail
We are consistently facing a suspend failure due to sighting alert
772439, impacting all projects of ADL-P, ADL-N, TGL, RPL-P,
RPL-P Refresh, and TWL. This issue interrupts tests, requiring log
collection, upload, and analysis, along with special coreboot builds to
mitigate the problem.

This is a waived issue. Waiver reason:
The issue can be recovered during the next suspend cycle. Since there
is a retry mechanism in end-user mode, it does not impact the end-user
experience.

To reduce issue complexity and accelerate development, this change
introduces an ignore check that allows the system to bypass the sighting
alert case. This enables the suspend process to proceed without
interruption by the waived case.

BUG=b:271527450, b:442730320
TEST=Run Dozing stress test and suspend_stress_test normally.

Change-Id: I93b5820fc092cf198eeceded06c9cf391d201873
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5968822
Tested-by: Curtis Chen <curtis.chen@intel.com>
Commit-Queue: Curtis Chen <curtis.chen@intel.com>
Reviewed-by: Eizan Miyamoto <eizan@chromium.org>
Reviewed-by: Stanisław Kardach <skardach@google.com>