hwdrm-videoproc-ta: Add few more parameters to the H264 slice header parse result.
The H264 parser now extracts a few more slice header info from the
bitstream. The `H264SliceHeaderData` and `StreamDataForSliceHeader`
structs are updated for the new variables.

To allow for a phased rollout, the TEE service is updated to be
backward compatible. It detects if the output is the older one by
looking at the size of the structure and execute the conversion
between the older and newer structures.

BUG=b:449845359
TEST=video.PlayDRM.cencv1_h264_ctr

Cq-Depend: chrome-internal:8652191
Change-Id: I7f3564450f2d56cefa8eeb2a229a3d5bee1ae64f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/7021615
Tested-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Nathan Hebert <nhebert@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>