chromeos-config: Add property to disable power button
Add a cros-config entry when the power button is on the keyboard.
It will be used to disable that button when in tablet mode, to
avoid unintentional power-off.
Add similar flag for chrome, which handle power button in S0.

BUG=b:422120500
TEST=On Jubilant, check the property is available:
cat /run/chromeos-config/v1/keyboard/disable-power-button-in-tablet-mode : true

Change-Id: Ia7a7fc4759836a60d8260615bdea8963a631fc4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6951957
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Aamir Bohra <aamirbohra@google.com>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>