kdump: Support kdump-dev on non-LVM devices
The current kdump-dev implementation relies on the fact that the
stateful partition is managed by LVM, which is not the case on older
CrOS devices. We need to find another way to save the crash dump to the
stateful partition in the kdump kernel when LVM is not used.

This patch adds another code path to mount the stateful partition and
save the dumps into /dev_image. dev_image is only available in dev mode
and will be mounted to /usr/local in a normal boot (i.e., not in the
kdump kernel).

Note that we still need to keep the LVM code path because when LVM is
used the stateful partition is not in the ext4 format and thus cannot be
mounted directly. We still need lvm for disk operation.

BUG=b:457988924
TEST=`tast run $DUT crash.Kdump` on brya and hatch.

Cq-Depend: chromium:7123320
Change-Id: I6636d7d869613f8d585eb9ebf8258a2e3d9766d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/7123665
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>