camera: Make header constant inline
The constant cros::CameraClientOps::kStreamId is defined in a header
file, but not marked as inline. This violates the One-Definition-Rule
and can lead to problems. In fact it does so in the current libchrome
uprev CL where Tast tests are failing as a consequence of the following
error:

2025-11-12T00:23:27.580437Z [00:23:26.969] Command: cros_camera_connector_test --gtest_filter=ConnectorTest.GetInfo
2025-11-12T00:23:27.581678Z [00:23:26.969] Uncaptured output:
cros_camera_connector_test: symbol lookup error: /usr/lib64/libcros_camera.so: undefined symbol: _ZN4cros15CameraClientOps9kStreamIdE

(None of the changes included in the uprev CL looks related.)

Making the constant `static inline constexpr` fixes the issue.

BUG=None
TEST=cros_camera_connector_test --gtest_filter=ConnectorTest.GetInfo

Change-Id: I26244e8c71276bb39705ba6ea640a7ae50516939
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/7145299
Tested-by: Georg Neis <neis@chromium.org>
Commit-Queue: Georg Neis <neis@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>