ml: Fix multiprocess tests with ipcz enabled
The `WebPlatformHandwritingModel.LoadModelAndRecognize` multiprocess
test was failing when built with ipcz. The failure was caused by an
incorrect Mojo IPC setup in the test environment, where the test runner
should act as its own broker. This conflicts with production code that
assumes it is a non-broker process sharing a system-wide broker, leading
to an invalid connection attempt.

BUG=chromium:41494930
TEST=CQ

Change-Id: I599cd82483b1de0428926f8415f8781f40d71f10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/7152233
Commit-Queue: Andrea Orru <andreaorru@chromium.org>
Reviewed-by: Xinglong (Alan) Luan <alanlxl@chromium.org>
Tested-by: Andrea Orru <andreaorru@chromium.org>