flexor: Replace use of sgdisk with pure Rust
The environment we're building mass deploy images in doesn't have
sgdisk, and due to time pressures this seems to be the most expedient
way to resolve the issue.

We try align partitions in the same way that sgdisk does. There's a
speed penalty when using a partition that isn't aligned to an
appropriate boundary, though I don't expect it to be a big difference:
* This image will be written to disk as a whole, ignoring partitions
  and so avoiding the speed penalty
* This should only boot one time with this layout, after which flexor
  will overwrite the whole disk
* By most accounts the speed cost is 10-30%
it seems borderline worth it to me, and comes down to how heavily you
weight bytes-over-the-wire and install efficiency.

BUG=b:445653477
TEST=$ cargo xtask create-flexor-disk --frd-bundle .../frd-debug/ \
          --crdyboot-verbose
     $ cargo xtask run-flexor-disk flexor_disk.img
TEST=$ cargo xtask create-flexor-disk --frd-bundle .../frd-debug/ \
          --crdyboot-verbose \
          --mass-deployable
     $ truncate --size=24GiB flexor_disk.img
     $ cargo xtask run-flexor-disk flexor_disk.img
TEST=Test on kokoro (by copying the xtask subdir to a test repo, with
     appropriate modifications to Cargo.toml and build scripts), see
     a successful build.

Change-Id: I711366ac44af2de6b1527a2d7bec70767ac3ce31
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/7168866
Auto-Submit: Ted Brandston <tbrandston@google.com>
Reviewed-by: Alexis Savery <asavery@chromium.org>
Tested-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Alexis Savery <asavery@chromium.org>