system-proxy: Allow kill with SIGPIPE signal on arm64
This change allows to perform `kill` syscall with SIGPIPE signal. This
prevents the outer process from crashing when the inner process
(the worker) triggers a `SIGPIPE` due to a forbidden syscall, and
allows the outer process to handle the error gracefully.

Note that the issue was only present on ARM64, thus patching seccomp
policy for that architecture only.

BUG=b:461319239
TEST=None

Change-Id: I016a490dd61618f94b519133adfe394bf99b7aaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/7168503
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Andrei Nedoluzhko <nedol@google.com>
Tested-by: Andrei Nedoluzhko <nedol@google.com>
Reviewed-by: Andreea Costinas <acostinas@google.com>