flexor: Write install_type file for metrics
Flexor is gaining the ability to pass along a file indicating the
install method, either mass-deploy or remote-deploy.

This xtask is used for creating the Mass Deploy images, and even for
Remote Deploy it's handy to have that for testing. Add the install_type
file to images made with create-flexor-image.

Also change the tempfile dep, to match what's available in rust_crates.

BUG=b:462074005
TEST=`xtask create-flexor-disk --frd-bundle ~/images/frd-2-M138-stable`
     with and without `--mass-deployable`, and then inspect the images
     with `guestfish`.
TEST=$ cargo test -p xtask

Change-Id: Ic9dbbc59006b9df3d7db22f552e3a3cd5620fad3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/7213627
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Ted Brandston <tbrandston@google.com>