cryptohome: Get rid of the pinweaver password feature flag
This has now been rolled out for quite a while and default enabled as
well, time to drop the code. This should make it slightly easier to do
some fixes for the feature since they don't need to be conditionally
gated.

The change is somewhat bigger than you'd expect because it also backs
out the Feature API from the auth factor driver manager, since that was
the only actual use of said API. It's mostly just stripping the
parameter out of the constructor though, and updating a bunch of the
constructor calls to no longer need to pass it in.

BUG=b:471804160
TEST=built package, ran unit tests

Change-Id: I1b82b9946543b2d5088e11d9e57fdd2e70dd5df9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/7224694
Reviewed-by: Zi Lin <lziest@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>