system-proxy: Pass ConnectedNamespace data to ProxyConnectJob
When root namespace feature is enabled for DNS proxy, system-proxy would
not be able to reach DNS proxy's address in /etc/resolv.conf as it is a
localhost address. This is because system-proxy lives in patchpanel's
ConnectedNamespace.

This patch passes ConnectedNamespace data to eventually be used in a
follow-up patch to query DNS proxy listening on the host-side address of
the ConnectNamespace.

BUG=b:363147822
TEST=manual with the DNS patch on the chain, observe system-proxy
     working with root namespace dns-proxy.

Change-Id: If7a42a3db198ad3e24dfa158ea66f0e80feb01be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/7226274
Commit-Queue: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Andreea Costinas <acostinas@google.com>
Tested-by: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>