system-proxy: Use ConnectNamespace host address for dns-proxy feature
When root namespace feature is enabled for DNS proxy, system-proxy would
not be able to reach DNS proxy's address in /etc/resolv.conf as it is a
localhost address. This is because system-proxy lives in patchpanel's
ConnectedNamespace.

This patch reads /etc/resolv.conf to see if the feature is enabled. If
so, libcurl is set to use the host address of ConnectNamespace where DNS
proxy is listening.

BUG=b:363147822
TEST=manual; see system-proxy working with root ns dns-proxy;
TEST=tast run network.SystemProxyForSystemServices;
TEST=manual; observe that dns search list is honored;

Change-Id: I8b3ae91c30642a667b5bf9e23d68c4d4fe8d444b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/7226275
Tested-by: Jason Iman <jasongustaman@chromium.org>
Commit-Queue: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>