modemfwd: Apply generic carrier in ForceFlash
In the case where the carrier partition is corrupted, a ForceFlash
initiated by the wedge check leaves the carrier uuid empty, so the
recovery tooling will never apply a correct carrier payload. This
changes updates ForceFlash to use a default "generic" carrier in this
scenario, to ensure that all payloads are re-applied when something is
wrong on the modem side.

BUG=b:430959577
TEST=Verified that a carrier payload is given to the helper during force-flash.

Change-Id: I4ea69ee867b9c238e1063f0d4f7dca8d13f94148
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/7228005
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>