cryptohome: Make time-limited the std policy for password factors
This implementation is not ideal, because it basically defaults to
time-limited for all passwords when an individual password might not be,
particularly if a machine does not support pinweaver.

However, that's not easily solvable because auth factor metadata is
defined without reference to an actual auth factor (e.g. with an auth
block) or hardware, which makes all the checks we'd have to do not
really possible. This is to some extent a limitation of how auth factor
metadata is structured. While we could do a much more massive refactor
of how the metadata is derived and reflected up to the API that's kind
of hard to justify; it seems better to just specify that the factor is
time-limited even if for some device there would never be an actual
non-zero delay.

BUG=b:461804160
TEST=built package, ran unit tests

Disallow-Looks-For-Green: True
Disallow-Recycled-Builds: test-failures
Change-Id: Ia20abaef423e3aaefb0a71e02413dce3084aa708
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/7228006
Reviewed-by: Zi Lin <lziest@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>