flex_bluetooth: Add retry for BT device detection
Sleep and retry if no BT devices are found. This will ensure the
flex overrides are properly applied.

The current state is to retry 10 times (11 total attempts) and to
sleep 5 seconds between each attempt.

Modifying this is ok, but note: too fast and the
overrides will not be applied; too slow and it will affect overall
BT stack startup time.

BUG=b:430623502
TEST=check logs on older Flex dut;
     observe override written in sysprop

Change-Id: Id2f2e3ddd005d59f57dc37bde78c3eb3ebfa1946
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/7266329
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Tested-by: Nehemiah Dureus <ndureus@chromium.org>
Reviewed-by: Hsin-chen Chuang <chharry@google.com>
Commit-Queue: Nehemiah Dureus <ndureus@chromium.org>