[owners] Remove yanivt@google.com from hermes/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/yanivt@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I89e9fdf4d2b7f47de0c4776e5ace73b89d42562b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/7317697
Tested-by: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Madhav <madhavadas@google.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Madhav <madhavadas@google.com>